Total code restructure
* Modularized classes
