+++ bower/commands/discover.py
@@ -1,0 +1,5 @@
+import sublime_plugin
+
+class DiscoverPackageCommand(sublime_plugin.WindowCommand):
+    def run(self):
+        self.window.run_command('open_url', {'url': 'http://sindresorhus.com/bower-components'})
