+++ bower/exceptions/non_clean_exit_error.py
@@ -1,0 +1,6 @@
+class NonCleanExitError(Exception):
+    def __init__(self, returncode):
+        self.returncode = returncode
+
+    def __str__(self):
+        return repr(self.returncode)
